fix and clear code

+++ MK4duo/src/HAL/HAL_DUE/HAL_Due.h
@@ -174,9 +174,7 @@
 #undef HIGH
 #define HIGH        1
 
-// intRes = intIn1 * intIn2 >> 16
 #define MultiU16X8toH16(intRes, charIn1, intIn2)   intRes = ((charIn1) * (intIn2)) >> 16
-// intRes = longIn1 * longIn2 >> 24
 #define MultiU32X32toH32(intRes, longIn1, longIn2) intRes = ((uint64_t)longIn1 * longIn2 + 0x80000000) >> 32
 // Macros for stepper.cpp
 #define HAL_MULTI_ACC(intRes, longIn1, longIn2) MultiU32X32toH32(intRes, longIn1, longIn2)
@@ -185,7 +183,7 @@
 
 // TEMPERATURE
 #undef analogInputToDigitalPin
-#define analogInputToDigitalPin(p) ((p < 12u) ? (p) + 54u : -1)
+#define analogInputToDigitalPin(p) ((p < 12) ? (p) + 54 : -1)
 // Bits of the ADC converter
 #define ANALOG_INPUT_BITS 12
 #define ANALOG_REDUCE_BITS 0
@@ -202,7 +200,7 @@
 // --------------------------------------------------------------------------
 
 typedef uint32_t millis_t;
-typedef uint8_t Pin;
+typedef int16_t Pin;
 
 // --------------------------------------------------------------------------
 // Public Variables
